pyproject.toml
[project]
name = "crawl4ai-mcp"
version = "0.1.0"
description = "MCP server for integrating web crawling and RAG into AI agents and AI coding assistants"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crawl4ai[pdf]==0.6.2",
    "mcp==1.7.1",
    "supabase==2.15.1",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=4.1.0",
    "neo4j>=5.28.1",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "protobuf>=3.20.0",
    "PyPDF2>=3.0.0",
    "accelerate>=1.9.0",
]
